Difficulty issues sorted (calling the same code twice... does nothing good). Needs the following: - PFP and image implementation onto flashcards - Match game needs sorting - Suggested decks looks ugly, and shows nothing if no decks - Summary screen does not show the actual information needed - Comments !!!!

+++ app/src/main/java/uk/ac/aber/dcs/cs31620/revisionmaster/model/dataclasses/Difficulty.kt
@@ -3,8 +3,8 @@ package uk.ac.aber.dcs.cs31620.revisionmaster.model.dataclasses
 /**
  * Enum class for the difficulty, including its label and weighting
  */
-enum class Difficulty(val label: String, val weight: Int) {
-    EASY("Easy", 1),
-    MEDIUM("Medium", 2),
-    HARD("Hard", 3)
+enum class Difficulty(val label: String, val weight: Double) {
+    EASY("Easy", 1.0),
+    MEDIUM("Medium", 2.0),
+    HARD("Hard", 3.0)
 }
